making shared easier to use from azure functions
