+++ src/web/Server/Extensions/AzureAppConfigurationExtensions.cs
@@ -4,7 +4,6 @@ public static class AzureAppConfigurationExtensions
 {
     public static WebApplicationBuilder AddAzureAppConfiguration(this WebApplicationBuilder builder)
     {
-
         string? connectionString = builder.Configuration.GetConnectionString("AppConfig");
         if (connectionString is null)
         {
@@ -15,7 +14,6 @@ public static class AzureAppConfigurationExtensions
 
         builder.Services.AddAzureAppConfiguration();
         builder.Configuration.AddAzureAppConfiguration(options =>
-        {
             options.Connect(connectionString)
                 // Load all keys that have no label
                 .Select("*")
@@ -25,8 +23,7 @@ public static class AzureAppConfigurationExtensions
                     refreshOptions.Register("Sentinel", refreshAll: true);
                     refreshOptions.SetCacheExpiration(TimeSpan.FromMinutes(5));
                 })
-                .UseFeatureFlags(flagOptions => flagOptions.CacheExpirationInterval = TimeSpan.FromMinutes(3));
-        });
+                .UseFeatureFlags(flagOptions => flagOptions.CacheExpirationInterval = TimeSpan.FromMinutes(3)));
 
         return builder;
     }
